xAImicroservices/SHAP/SHAP-Visualization/requirements.txt
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.32.0
matplotlib==3.7.1
matplotlib-inline==0.1.6
pydotplus==2.0.2
cloudpickle==2.2.1
joblib==1.2.0
llvmlite==0.39.1
numba==0.56.4
numpy==1.23.5
packaging==23.0
pandas==1.5.3
python-dateutil==2.8.2
pytz==2023.3
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
slicer==0.0.7
threadpoolctl==3.1.0
tqdm==4.65.0
typing_extensions==4.5.0
dill==0.3.6
shap==0.41.0
requests==2.28.2
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
keras==2.12.0

pymongo==4.3.3
confluent-kafka==2.1.1
paho-mqtt==1.6.1
